requirements.txt
python-barcode
PIL